Move GeoJson utility
