+++ src/pages/search/dashboard/landscape/humanFootprint/TimelineFootprintController.ts
@@ -4,8 +4,7 @@ import { shapeLayer } from "pages/search/types/layers";
 import matchColor from "pages/search/utils/matchColor";
 import { ShapeAPIObject } from "pages/search/types/api";
 import { CancelTokenSource } from "axios";
-import { hasValidGeoJSONData } from "utils/GeoJsonUtils";
-import { AnyARecord } from "node:dns";
+import { hasValidGeoJSONData } from "pages/search/utils/GeoJsonUtils";
 
 type SEKeys = Record<"paramo" | "dryForest" | "wetland", string>;
